# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dispatch controller API
  version: 0.1.0
  description: Project-scoped application deployment and operations API.
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /apps/{id}/observations:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      operationId: getApplicationObservations
      description: Requires project.view. Read configuration, separate workload and endpoint observations, freshness,
        and delivery history. This read does not contact the target.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationStatus'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    put:
      operationId: configureApplicationObservations
      description: Requires project.configure. Scheduled observations and notifications are opt-in. Revision guards
        concurrent edits; sensitive delivery URLs are write-only. Stale threshold must be at least the interval.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationStatus'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationConfigurationInput'
  /apps/{id}/observations/check:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    post:
      operationId: checkApplicationObservations
      description: Requires project.configure. Manual bounded workload and HTTP/TLS check from Dispatch controller;
        results are observations of that location only. Active runtime operations return conflict.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationStatus'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ObservationConfiguration:
      type: object
      properties:
        appId:
          type: string
        projectId:
          type: string
        revision:
          type: integer
        scheduled:
          type: boolean
        intervalSeconds:
          type: integer
          minimum: 60
          maximum: 86400
        staleAfterSeconds:
          type: integer
          minimum: 60
          maximum: 604800
        endpointUrl:
          type: string
        notificationsEnabled:
          type: boolean
        webhookConfigured:
          type: boolean
        mutedUntil:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    ObservationConfigurationInput:
      type: object
      required:
      - revision
      - scheduled
      - intervalSeconds
      - staleAfterSeconds
      - notificationsEnabled
      properties:
        revision:
          type: integer
        scheduled:
          type: boolean
        intervalSeconds:
          type: integer
          minimum: 60
          maximum: 86400
        staleAfterSeconds:
          type: integer
          minimum: 60
          maximum: 604800
        endpointUrl:
          type: string
          description: Optional HTTP(S) endpoint without user information or query credentials.
        notificationsEnabled:
          type: boolean
        webhookUrl:
          type: string
          writeOnly: true
          description: Optional HTTPS delivery URL; omitted preserves the existing encrypted destination.
        removeWebhook:
          type: boolean
        mutedUntil:
          type:
          - string
          - 'null'
          format: date-time
    EndpointObservation:
      type: object
      properties:
        state:
          type: string
        tls:
          type: string
        httpStatus:
          type: integer
        certificateExpiresAt:
          type: string
          format: date-time
        durationMs:
          type: integer
        message:
          type: string
        location:
          type: string
          const: Dispatch controller
    ApplicationObservation:
      type: object
      properties:
        appId:
          type: string
        projectId:
          type: string
        configurationRevision:
          type: integer
        deploymentId:
          type: string
        source:
          type: string
        state:
          type: string
        drift:
          type: string
        health:
          type: string
        endpoint:
          $ref: '#/components/schemas/EndpointObservation'
        checkedAt:
          type: string
          format: date-time
        nextCheckAt:
          type: string
          format: date-time
        consecutiveFailures:
          type: integer
        message:
          type: string
        location:
          type: string
    ObservationEvent:
      type: object
      properties:
        id:
          type: string
        appId:
          type: string
        projectId:
          type: string
        deploymentId:
          type: string
        configurationRevision:
          type: integer
        kind:
          type: string
        previousState:
          type: string
        state:
          type: string
        message:
          type: string
        link:
          type: string
        createdAt:
          type: string
          format: date-time
        delivery:
          type: string
        attempts:
          type: integer
        nextAttemptAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
        deliveryMessage:
          type: string
    ObservationStatus:
      type: object
      properties:
        configuration:
          $ref: '#/components/schemas/ObservationConfiguration'
        observation:
          $ref: '#/components/schemas/ApplicationObservation'
        freshness:
          type: string
          enum:
          - not_checked
          - fresh
          - stale
          - checking
        checking:
          type: boolean
        events:
          type: array
          items:
            $ref: '#/components/schemas/ObservationEvent'
...
    Problem:
      type: object
      required: [status, title, detail]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }